Nuclide object, branching & linear chains, diagram
The main branch has been refactored to simplify the library. We are
sticking with a single Nuclide class which defines a nuclide and has
access to properties such as calculating it's decay chains etc. Nuclide
objects contain all of the information necessary for the nuclide that
was asked for by taking a subset of the nubase database. Decay chains
are calculated using the BR column of nubase, and a decay key in the
database config file. Sucsessive nuclides in a decay chain all reference
is the (original) nubase dataframe, therefore reducing the load on RAM
to 1 instance of the databases. The way it has been written currently is
that a user can explicitly input a nubase database, otherwise it will
search through the databases folder for .txt files with the name
nubase{year}.txt. It currently only searches for nubase2020 and
nubase2016, in that order.